+++ src/main/java/org/example/vehicle/factory/CarFactory.java
@@ -7,7 +7,7 @@ import org.example.vehicle.Vehicle;
 public final class CarFactory {
 
     private static final String DEFAULT_ENGINE = "1.6L I4";
-    private static final float DEFAULT_SEATS = 5;
+    private static final int DEFAULT_SEATS = 5;
     private static final float PRICE_PER_DAY = 5000;
     private static final float PRICE_PER_HOUR = 650;
 
@@ -25,7 +25,7 @@ public final class CarFactory {
     }
 
     public static Vehicle createCustomCar(String id, String model, String engine,
-                                          float seats, float pricePerDay, float pricePerHour) {
+                                          int seats, float pricePerDay, float pricePerHour) {
         return new Car(id, model, engine, seats, pricePerDay, pricePerHour);
     }
 }
